Rename modal files
